refactor(Breadcrumbs): rewrite stories
