+++ src/components/Breadcrumbs/__stories__/OnClick.stories.tsx
@@ -1,0 +1,28 @@
+import { Story } from '@storybook/react'
+import { ComponentProps, useState } from 'react'
+import Breadcrumbs from '..'
+
+export const OnClick: Story<ComponentProps<typeof Breadcrumbs>> = () => {
+  const [value, setValue] = useState(1)
+
+  return (
+    <Breadcrumbs variant="bubble" selected={value}>
+      <Breadcrumbs.Item onClick={(_, step) => setValue(step)}>
+        Step 1
+      </Breadcrumbs.Item>
+      <Breadcrumbs.Item onClick={(_, step) => setValue(step)}>
+        Step 2
+      </Breadcrumbs.Item>
+      <Breadcrumbs.Item disabled onClick={(_, step) => setValue(step)}>
+        Step 3
+      </Breadcrumbs.Item>
+    </Breadcrumbs>
+  )
+}
+
+OnClick.parameters = {
+  docs: {
+    storyDescription:
+      'You can make `Breadcrumbs.Item` clickable with `onClick` handler which pass `(event, stepClicked)` params. You can also disabled the onClick handler by using `disabled` prop',
+  },
+}
